add main process

+++ models/metric.py
@@ -1,8 +1,8 @@
-from sqlalchemy import Column, String, BigInteger
+from sqlalchemy import Column, String, Float
 from models.base import BaseDatabaseModel
 
 class Metric(BaseDatabaseModel):
     __tablename__ = 'metric'
 
     function_name = Column(String, default="")
-    execution_time = Column(BigInteger, default="")
+    execution_time = Column(Float, default="")
